requirements.txt
Django==3.2.7
djangorestframework==3.12.4
beautifulsoup4==4.10.0
requests==2.26.0
